# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fintech PWA API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /cards:
    get:
      operationId: listCards
      tags: [cards]
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
    post:
      operationId: issueCard
      tags: [cards]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardInput"
      responses:
        "201":
          description: Card issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}:
    get:
      operationId: getCard
      tags: [cards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}/topup:
    post:
      operationId: topupCard
      tags: [cards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupInput"
      responses:
        "200":
          description: Topup initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /cards/{id}/withdraw:
    post:
      operationId: withdrawCard
      tags: [cards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawInput"
      responses:
        "200":
          description: Withdraw initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # TRANSACTIONS
...
  /gift-cards:
    get:
      operationId: listGiftCards
      tags: [giftcards]
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: country
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Gift cards catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GiftCard"
  /gift-cards/{id}:
    get:
      operationId: getGiftCard
      tags: [giftcards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gift card details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftCard"
...
  /games:
    get:
      operationId: listGames
      tags: [games]
      parameters:
        - name: platform
          in: query
          schema:
            type: ["string", "null"]
        - name: region
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Games catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"

  # GAME TOP-UP SERVICES
...
  /game-services/{id}/topup:
    post:
      operationId: topupGameService
      tags: [games]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameTopupInput"
      responses:
        "200":
          description: Top-up initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ESIM
  /esim-plans:
    get:
      operationId: listEsimPlans
      tags: [esim]
      parameters:
        - name: tab
          in: query
          schema:
            type: ["string", "null"]
            enum: ["countries", "regions", "global", null]
        - name: country
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: eSIM plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EsimPlan"
  /esim-plans/{id}/purchase:
    post:
      operationId: purchaseEsim
      tags: [esim]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # TRANSFERS
...
  /mobile-operators:
    get:
      operationId: listMobileOperators
      tags: [mobile]
      parameters:
        - name: country
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Mobile operators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MobileOperator"
  /mobile-topup:
    post:
      operationId: mobileTopup
      tags: [mobile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileTopupInput"
      responses:
        "200":
          description: Top-up order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # REVIEWS
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        telegramId:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, email, createdAt]
    RegisterInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required: [email, password]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    Card:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [virtual, plastic]
        system:
          type: string
          enum: [visa, mastercard]
        numberMasked:
          type: string
        balance:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [active, blocked, pending]
        createdAt:
          type: string
      required: [id, userId, type, system, numberMasked, balance, currency, status, createdAt]
    CardInput:
      type: object
      properties:
        type:
          type: string
          enum: [virtual, plastic]
        system:
          type: string
          enum: [visa, mastercard]
        currency:
          type: string
      required: [type, system]
    TopupInput:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [card, usdt, sbp, cash]
      required: [amount, method]
    WithdrawInput:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [card, usdt, sbp]
        destination:
          type: string
      required: [amount, method]
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        cardId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [topup, withdraw, payment]
        method:
          type: string
          enum: [card, usdt, sbp, cash]
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, success, failed]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, type, method, amount, currency, status, createdAt]
    GiftCard:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        country:
          type: string
        countryFlag:
          type: string
        denominations:
          type: array
          items:
            type: number
        image:
          type: string
        priceRub:
          type: number
        popular:
          type: boolean
      required: [id, name, category, country, countryFlag, denominations, image, priceRub, popular]
...
    Game:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        platform:
          type: string
          enum: [steam, xbox, playstation, nintendo, other]
        region:
          type: string
        priceRub:
          type: number
        image:
          type: string
        description:
          type: string
        genre:
          type: string
        popular:
          type: boolean
      required: [id, name, platform, region, priceRub, image, description, genre, popular]
...
    GameTopupInput:
      type: object
      properties:
        accountId:
          type: string
        amount:
          type: number
      required: [amount]
    EsimPlan:
      type: object
      properties:
        id:
          type: integer
        country:
          type: ["string", "null"]
        countryFlag:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        dataGb:
          type: number
        days:
          type: integer
        priceRub:
          type: number
        tab:
          type: string
          enum: [countries, regions, global]
        popular:
          type: boolean
      required: [id, dataGb, days, priceRub, tab, popular]
...
    MobileOperator:
      type: object
      properties:
        id:
          type: integer
        country:
          type: string
        countryFlag:
          type: string
        operator:
          type: string
        logo:
          type: string
        amounts:
          type: array
          items:
            type: number
      required: [id, country, countryFlag, operator, logo, amounts]
    MobileTopupInput:
      type: object
      properties:
        operatorId:
          type: integer
        phoneNumber:
          type: string
        amount:
          type: number
      required: [operatorId, phoneNumber, amount]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        productId:
          type: ["integer", "null"]
        productName:
          type: string
        productType:
          type: string
          enum: [gift_card, game, game_topup, esim, mobile_topup, transfer]
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, success, failed]
        createdAt:
          type: string
      required: [id, productName, productType, amount, currency, status, createdAt]
...