# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fintech PWA API
 * OpenAPI spec version: 0.1.0
 */
import type { GamePlatform } from './gamePlatform';

export interface Game {
  id: number;
  name: string;
  platform: GamePlatform;
  region: string;
  priceRub: number;
  image: string;
  description: string;
  genre: string;
  popular: boolean;
}